Create metadata-store-directory-common module (#771)
metadata-store-directory-common module was created to avoid circular dependency because the RoutingData classes will be used in zookeeper-api.
